Adds functionality for items and people. Shores up input as well.
Still need to make it so the program doesn't flip out when you enter a
command that isn't possible/implemented yet. Also, a save and quit function, but it's in a safe state to start adding rooms and people and objects. Objects don't do anything yet. Need to figure out how to make them perform actions cleanly.